+++ Samples/iOS-Swift/iOS-Swift/AppDelegate.swift
@@ -1,6 +1,9 @@
 import SentrySampleShared
 import UIKit
 
+import Logging
+import SentrySwiftLog
+
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
     private var randomDistributionTimer: Timer?
@@ -24,10 +27,21 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         SentrySDKWrapper.shared.startSentry()
         SampleAppDebugMenu.shared.display()
-
+        
         if #available(iOS 15.0, *) {
             metricKit.receiveReports()
         }
+        
+        // Use structured logs with swift-log
+        LoggingSystem.bootstrap { _ in
+            return SentryLogHandler(logLevel: .trace)
+        }
+        let logger = Logger(label: "io.sentry.iOS-Swift")
+        logger.trace(
+            "[iOS-Swift] didFinishLaunchingWithOptions",
+            metadata: ["foo": "bar"],
+            source: "iOS"
+        )
         
         return true
     }